[fix] Moved code according to structure
