+++ src/crypto/AddressInterface.js
@@ -1,0 +1,6 @@
+// @flow
+
+export interface YoroiAddressInfoInterface<KeyHashes> {
+  +address: string;
+  +getKeyHashes: () => Promise<KeyHashes | null>;
+}
